Cleanup warnings (#6)
1. adding a global definition for Windows to suppress deprecated functions warnings.
2. Cleanup a bunch of warning mainly by setting explicit casting:
  - explicit cast on signed/unsigned.
  - explicit cast when reducing format (i.e double to float).
3. removing return value on function defined `void function()`
4. renaming Windows i/o files functions.

+++ libinstpatch/IpatchDLS2.c
@@ -433,7 +433,7 @@ ipatch_dls2_base_find_unused_locale (IpatchBase *base, int *bank,
   GSList *locale_list = NULL;
   IpatchDLS2Inst *inst;
   GSList *p;
-  int b, n;			/* Stores current bank and preset number */
+  guint32 b, n;			/* Stores current bank and preset number */
   guint32 lbank, lprogram;
 
   /* fill array with bank and preset numbers */
@@ -459,8 +459,8 @@ ipatch_dls2_base_find_unused_locale (IpatchBase *base, int *bank,
 
   locale_list = g_slist_sort (locale_list, (GCompareFunc)locale_gcompare_func);
 
-  b = *bank;
-  n = *program;
+  b = (guint32)*bank;
+  n = (guint32)*program;
 
   /* loop through sorted list of bank:programs */
   p = locale_list;
