+++ libinstpatch/IpatchSF2VoiceCache.c
@@ -124,7 +124,7 @@ ipatch_sf2_voice_cache_finalize (GObject *gobject)
 {
   IpatchSF2VoiceCache *cache = IPATCH_SF2_VOICE_CACHE (gobject);
   IpatchSF2Voice *voice;
-  int i;
+  guint i;
 
   g_free (cache->sel_info);
 
@@ -305,7 +305,7 @@ ipatch_sf2_voice_cache_set_voice_range (IpatchSF2VoiceCache *cache,
 
   g_return_if_fail (IPATCH_IS_SF2_VOICE_CACHE (cache));
   g_return_if_fail (voice != NULL);
-  g_return_if_fail (sel_index < cache->sel_count);
+  g_return_if_fail (sel_index < (guint)cache->sel_count);
   g_return_if_fail (low <= high);
 
   rangep = (int *)(cache->ranges->data);
